packages/jumpstarter-robotframework/pyproject.toml
[project]
name = "jumpstarter-robotframework"
dynamic = ["version", "urls"]
description = ""
authors = [{ name = "Kirk Brauer", email = "kbrauer@hatci.com" }]
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = ["jumpstarter", "robotframework>=7.2.0"]

[dependency-groups]
dev = ["jumpstarter-driver-power", "robotframework-tidy>=4.0.0"]

[tool.hatch.metadata.hooks.vcs.urls]
Homepage = "https://jumpstarter.dev"
source_archive = "https://github.com/jumpstarter-dev/repo/archive/{commit_hash}.zip"

[tool.hatch.version]
source = "vcs"
raw-options = { 'root' = '../../' }

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"
